Made LoginTask class static to prevent potential memory leaks.
Non static inner classes hold a reference to the parent class. This can prevent the parent activity from being garbage collected. In the current design of ManagerLoginActivity, this will not happen because the task has to finish for the activity to stop being active, but let's be on the safe side.