Shifted functionality of GuestEventsFragment to GuestScheduleFragment.
The GuestEventsFragment right now does what the GuestScheduleFragment is supposed to do. This commit copies the functionality of the events fragment to the schedule fragment.

This change is the first of two changes. The next one will replace the functionality of the events fragment to what it is supposed to do: just display the name of the events, not their time/location.