Fixed a crash in ManagerSpeakerFragment on newly created events.
The events that are created through the overflow menu option do not add a corresponding speaker row in the database. This was causing a crash when trying to retrieve speaker details because we assumed the speaker would be present. Now, we insert a blank speaker row when none is found.