Moved the overflow options menu handling to fragments.
Earlier, ManagerActivity was handling the overflow menu options. Now, we move it to the individual fragments to handle all the things.

Also added a method in ManagerActivity to log out. It simply set the manager Id to 0. We did not need this earlier because the log out was being handled in ManagerActivity so there was no need to create a separate method.