Moved the creation of dummy speaker from speaker to events fragment.
The events fragment is a better place for the insertion because it adds the dummy as soon as the head logs in for the first time.

Also, this fixes a crash because we were trying to delete a speaker that did not exist. This happened when an event head tries to delete an event he created before ever opening the speaker fragment.